Remove python3-systemd-cython from debian dependencies
As it turns out, the 'official' systemd-python binding have a Logger
object as well. We can support both. (This may be a recent addition, as
I don't recall seeing this before.)

Since I'm having some packaging tribulations with the non-official
systemd-python package (see [1]), I'll opt for this solution for now.

[1]: https://github.com/mosquito/python-systemd/issues/4