Use dbusmock for test cases
This allows for more code coverage, because we can also check
the code paths that depend on getting configuration from DBus.
Most interestingly, this allows testing the javascript parsing
stuff in pacparser.

For now, the test cases only check that the curl commands work;
not whether they get the right data through the right path. I'm
not even sure if they do! A nice followup would be to have a proxy
return "HI THERE FROM THIS PROXY" and validate that curl finds that
proxy as it should.