runtests.py: add fake proxies to check proper settings discovery
This commit adds two proxies that respond with a distinct
string instead of the 'real' response to the request. This allows
validating that the proper proxy is being contacted.

In addition, this commit fixes a bug that was laid bare by
this: the Dhcp4Config property was added to the wrong object.