Support --force-proxy parameter
I couldn't find a way around having to clone a copy of forced_proxy
for every nested closure. Originally, I was hoping to just pass a
find_proxy closure to run_server, which separates concerns a bit nicer.
But there was absolutely no way to make the borrow checker happy.