Implement my own two_way_pipe so I can handle (dis)connection logic
Before this commit, pac4cli-rust would just hold on to all sockets
indefinitely.

There's still issues: the current logic doesn't guarantee that anything
that's in the buffer will still be sent downstream if upstream closes
the connection. Also, this way, we will *always* end with an error
condition, even if at the http level, we're good.