Be more precise about exception handling
This also removes an odd use of @inlineCallbacks for a function
that doesn't take any asynchronous actions.

It also moves the check for valid argument to outside of the get_config_wpad_url
function -- if it is not a valid argument, I'd rather raise an exception
than pass silently. Similarly, I remove the check for isfile according
to "rather ask for forgiveness than permission".