Create a Docker container with systemd
by running `test/test-arch-package.sh` a docker image with systemd init
running is built and a container is run.

The building of the image in itself is testing if the PKGBUILD is
successful in describing the build/install for `makepkg`.

The container is running pac4cli as a systemd service.

Next step is running tests inside on a native installation.