Remove listening to SIGHUP
I kept it thinking someone might still want to use it to integrate with
other network managers than NetworkManager (even if we don't ship that
integration ourselves). The pattern of reload-on-SIGHUP is pretty
typical, e.g. I think nginx does it too. (It actually reloads all
configuration, so by that logic, we should also reload /etc/pac4cli/*.)

But I now see my reasoning is flawed: there's no sense in "integrating
with other network managers" if the action that we're going to take on
SIGHUP is "talk to NetworkManager" :)

Thanks to @kdehairy for noticing this.