systemd config: wait for NetworkManager
Before this commit, we would start more-or-less simultaneously with
NetworkManager, as we were also part of the network target. This leads
to race conditions, and we'll terminate with an error if we cannot talk
to NetworkManager's dbus interface.

This was an issue for me at the reboot immediately following the upgrade
to Ubuntu 18.10, but I'm not sure it's related to that.

In anycase, it's probably better not to start unless NetworkManager is
up.