systemd service configuration: depend on NetworkManager
We'll need NetworkManager to start before us, so we can
sniff our network environment from it.

The systemd.unit(5) manpage describes the combination of
`Requires=` and `After=` as a typical way to describe
that we both (1) need NetworkManager to be enabled if
we are enabled and (2) if both units are enabled,
NetworkManager needs to start first.