systemd: make READY notifications work
Sending a READY=1 notification (as in pac4cli/servicemanager.py)
has no use unless we also configure `Type=notify`. And because
we fork out from the executable `pac4cli` to `python3 -m pac4cli`,
we need to configure `NotifyAccess=all` to ensure to child
process gets to send this notification.